04_advanced_rag/01_advanced_chunking_techniques/implementations/pdf_visual_summarizer/requirements.txt
pymupdf
langchain-core
langchain-openai
langchain-postgres
python-dotenv